requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
filetype==1.2.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-ollama==0.2.3
langchain-openai==0.3.8
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
Markdown==3.7
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.2.3
ollama==0.4.7
openai==1.66.3
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
zstandard==0.23.0
coloredlogs
langchain~=0.3.21
mcp~=1.4.1
python-docx~=1.1.2
uvicorn~=0.34.0
starlette~=0.46.1
pillow~=11.1.0
matplotlib~=3.10.1
seaborn~=0.13.2
typer~=0.15.2
fastapi~=0.115.12
markdown2~=2.5.3
langchain-experimental~=0.3.4
langchain-community~=0.3.20
aiosqlite~=0.21.0
langchain_mcp_adapters~=0.0.6
python-multipart~=0.0.20